+++ packages/lexical_mention_flutter/lib/lexical_mention_flutter.dart
@@ -42,4 +42,5 @@ export 'src/mention_scope.dart'
         MentionEditableBuilder,
         MentionItemBuilder,
         MentionScope,
-        MentionScopeState;
+        MentionScopeState,
+        MentionSurfaceBuilder;